Mutualize ConsumerStreamObserver classes
